requirements.txt
amqp==5.3.1
asgiref==3.11.1
asttokens==3.0.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
decorator==5.2.1
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-extensions==4.1
djangorestframework==3.17.0
executing==2.2.1
gunicorn==25.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.6.2
matplotlib-inline==0.2.1
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pre-commit==4.5.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
redis==7.3.0
ruff==0.15.7
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
